Revert "Add ability to automerge PRs based on labels"
