+++ .gitignore
@@ -114,6 +114,4 @@ terraform/terraform.tfstate.backup
 
 # virtualenv
 v-env/
-
-# VSCode
-.vscode
+build-env/
