+++ src/github/client.py
@@ -1,17 +1,12 @@
-from github import Github, PullRequest  # type: ignore
+from github import Github  # type: ignore
 from src.config import GITHUB_API_KEY
 
 gh_client = Github(GITHUB_API_KEY)
 
 
-def _get_pull_request(owner: str, repository: str, number: int) -> PullRequest:
+def edit_pr_description(owner: str, repository: str, number: int, description: str):
     repo = gh_client.get_repo(f"{owner}/{repository}")
     pr = repo.get_pull(number)
-    return pr
-
-
-def edit_pr_description(owner: str, repository: str, number: int, description: str):
-    pr = _get_pull_request(owner, repository, number)
     pr.edit(body=description)
 
 
@@ -21,8 +16,3 @@ def set_pull_request_assignee(owner: str, repository: str, number: int, assignee
     # allows you to *add* an assignee, not set the assignee.
     pr = repo.get_issue(number)
     pr.edit(assignee=assignee)
-
-
-def merge_pull_request(owner: str, repository: str, number: int, title: str, body: str):
-    pr = _get_pull_request(owner, repository, number)
-    pr.merge(commit_title=title, commit_message=body, merge_method="squash")
