+++ src/github/graphql/fragments/FullPullRequest.py
@@ -14,7 +14,6 @@ fragment FullPullRequest on PullRequest {
   closed
   merged
   mergedAt
-  mergeable
   url
   number
   repository {
@@ -71,11 +70,6 @@ fragment FullPullRequest on PullRequest {
           state
         }
       }
-    }
-  }
-  labels(last: 20) {
-    nodes {
-      name
     }
   }
 }
