+++ src/github/models/__init__.py
@@ -1,12 +1,10 @@
 from typing import Dict, Any
-from .pull_request import PullRequest, MergeableState
+from .pull_request import PullRequest
 from .review import Review, ReviewState
 from .comment import Comment
 from .issue_comment import IssueComment
 from .pull_request_review_comment import PullRequestReviewComment
 from .user import User
-from .commit import Commit
-from .label import Label
 
 
 def comment_factory(raw: Dict[str, Any]) -> Comment:
