+++ src/github/models/pull_request.py
@@ -1,25 +1,13 @@
 from datetime import datetime
-from typing import List, Optional, Dict, Any, Set
+from typing import List, Optional, Dict, Any
 from src.logger import logger
 from src.utils import parse_date_string
-from enum import Enum, unique
 
 # from .review import Review
 from .issue_comment import IssueComment
 from .review import Review
-from .commit import Commit
 from .user import User
-from .label import Label
 import copy
-
-
-@unique
-class MergeableState(Enum):
-    """https://developer.github.com/v4/enum/mergeablestate/"""
-
-    CONFLICTING = "CONFLICTING"
-    MERGEABLE = "MERGEABLE"
-    UNKNOWN = "UNKNOWN"
 
 
 class PullRequest(object):
@@ -81,7 +69,7 @@ class PullRequest(object):
     def number(self) -> int:
         return self._raw["number"]
 
-    def title(self) -> str:
+    def title(self) -> int:
         return self._raw["title"]
 
     def url(self) -> str:
@@ -121,37 +109,6 @@ class PullRequest(object):
     def merged(self) -> bool:
         return self._raw["merged"]
 
-    def mergeable(self) -> MergeableState:
-        return MergeableState(self._raw["mergeable"])
-
-    def is_mergeable(self) -> bool:
-        return self.mergeable() == MergeableState.MERGEABLE
-
-    # A PR is considered approved if it has at least one approval and every time changes were requested by a reviewer
-    # that same reviewer later approved.
-    def is_approved(self) -> bool:
-        if len(self.reviews()) > 0:
-            reviews = self.reviews()
-            reviews.sort(key=lambda review: review.submitted_at())
-
-            approved = False
-            reviewers_requesting_changes: Set[str] = set()
-            for review in reviews:
-                author_handle = review.author_handle()
-                if review.is_approval():
-                    approved = True
-                    if author_handle in reviewers_requesting_changes:
-                        reviewers_requesting_changes.remove(author_handle)
-                if review.is_changes_requested():
-                    reviewers_requesting_changes.add(author_handle)
-
-            return approved and len(reviewers_requesting_changes) == 0
-        else:
-            return False
-
-    def is_build_successful(self) -> bool:
-        return self.build_status() == Commit.BUILD_SUCCESSFUL
-
     def merged_at(self) -> Optional[datetime]:
         merged_at = self._raw.get("mergedAt", None)
         if merged_at is None:
@@ -168,10 +125,5 @@ class PullRequest(object):
         return copy.deepcopy(self._raw)
 
     def build_status(self) -> Optional[str]:
-        return self.commits()[0].status()
-
-    def commits(self) -> List[Commit]:
-        return [Commit(commit) for commit in self._raw["commits"]["nodes"]]
-
-    def labels(self) -> List[Label]:
-        return [Label(label) for label in self._raw["labels"]["nodes"]]
+        commit = self._raw["commits"]["nodes"][0]["commit"]
+        return commit["status"]["state"] if commit["status"] else None
