+++ test/impl/builders/__init__.py
@@ -3,8 +3,6 @@ from .comment_builder import CommentBuilder
 from .pull_request_builder import PullRequestBuilder
 from .review_builder import ReviewBuilder
 from .user_builder import UserBuilder
-from .commit_builder import CommitBuilder
-from .label_builder import LabelBuilder
 
 
 class Builder(object):
@@ -19,12 +17,6 @@ class Builder(object):
 
     def user(self, *args, **keywords):
         return UserBuilder(*args, **keywords)
-
-    def commit(self, *args, **keywords):
-        return CommitBuilder(*args, **keywords)
-
-    def label(self, *args, **keywords):
-        return LabelBuilder(*args, **keywords)
 
 
 def build(builder: BuilderBaseClass):
