+++ test/impl/builders/review_builder.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Union
+from typing import List, Union, Tuple, Optional, Dict, Any
 from datetime import datetime
 from .helpers import transform_datetime, create_uuid
 from src.github.models import Comment, Review, ReviewState, User
